// ui/vite-opentask/src/apis-gen/models/TaskInfo.ts
export interface TaskInfo {
    /**
     * 
     * @type {string}
     * @memberof TaskInfo
     */
    id?: string;
    /**
     * 
     * @type {string}
     * @memberof TaskInfo
     */
    appid?: string;
    /**
     * 
     * @type {string}
     * @memberof TaskInfo
     */
    name?: string;
    /**
     * 
     * @type {string}
     * @memberof TaskInfo
     */
    handler?: string;
    /**
     * 
     * @type {string}
     * @memberof TaskInfo
     */
    description?: string;
    /**
     * 
     * @type {string}
     * @memberof TaskInfo
     */
    timeType?: string;
    /**
     * 
     * @type {string}
     * @memberof TaskInfo
     */
    timeConf?: string;
    /**
     * 
     * @type {number}
     * @memberof TaskInfo
     */
    attemptInterval?: number;
    /**
     * 
     * @type {number}
     * @memberof TaskInfo
     */
    attemptMax?: number;
    /**
     * 
     * @type {string}
     * @memberof TaskInfo
     */
    handleParams?: string;
    /**
     * 
     * @type {string}
     * @memberof TaskInfo
     */
    scheduleMode?: string;
    /**
     * 
     * @type {string}
     * @memberof TaskInfo
     */
    alarmType?: string;
    /**
     * 
     * @type {string}
     * @memberof TaskInfo
     */
    alarmContent?: string;
    /**
     * 
     * @type {boolean}
     * @memberof TaskInfo
     */
    enabled?: boolean;
    /**
     * 
     * @type {string}
     * @memberof TaskInfo
     */
    triggerNextTime?: string;
    /**
     * 
     * @type {number}
     * @memberof TaskInfo
     */
    slot?: number;
}